requirements.txt
streamlit==1.32.0
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
marshmallow==3.20.1
python-dateutil==2.8.2 